refactor(#72): move interactions to directory
