+++ src/index.ts
@@ -1,3 +1,3 @@
 export { GoogleAnalytics } from "./GoogleAnalytics";
 export { usePagesViews } from "./usePagesViews";
-export { pageView, event } from "./gtag";
+export { pageView, event } from "./interactions";
