+++ src/interactions/pageView.test.ts
@@ -1,4 +1,4 @@
-import { event, pageView } from "./gtag";
+import { pageView } from "./pageView";
 
 const OLD_ENV = process.env;
 const mockGaMeasurementId = "mock";
@@ -14,83 +14,6 @@ beforeEach(() => {
 
 afterAll(() => {
   process.env = OLD_ENV;
-});
-
-describe("event", () => {
-  const mockEvent = "mock event";
-  const mockCategory = "mock category";
-  const mockLabel = "mock label";
-  const mockValue = 1;
-  const mockNonInteraction = true;
-
-  it("should not call gtag if measurement id is not set", () => {
-    process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID = undefined;
-
-    event(mockEvent);
-
-    expect(window.gtag).not.toBeCalled();
-  });
-
-  describe("options", () => {
-    process.env.NEXT_PUBLIC_GA_MEASUREMENT_ID = mockGaMeasurementId;
-
-    it("should call gtag with all the options", () => {
-      event(mockEvent, {
-        category: mockCategory,
-        label: mockLabel,
-        value: mockValue,
-        nonInteraction: mockNonInteraction,
-      });
-
-      expect(window.gtag).toBeCalledTimes(1);
-      expect(window.gtag).toHaveBeenCalledWith("event", mockEvent, {
-        event_category: mockCategory,
-        event_label: mockLabel,
-        value: mockValue,
-        non_interaction: mockNonInteraction,
-      });
-    });
-
-    it("should call gtag with {} when no options given", () => {
-      event(mockEvent);
-
-      expect(window.gtag).toBeCalledTimes(1);
-      expect(window.gtag).toHaveBeenCalledWith("event", mockEvent, {});
-    });
-
-    it("should call gtag with event_category when category given", () => {
-      event(mockEvent, {
-        category: mockCategory,
-      });
-
-      expect(window.gtag).toBeCalledTimes(1);
-      expect(window.gtag).toHaveBeenCalledWith("event", mockEvent, {
-        event_category: mockCategory,
-      });
-    });
-
-    it("should call gtag with event_label when event_label given", () => {
-      event(mockEvent, {
-        label: mockLabel,
-      });
-
-      expect(window.gtag).toBeCalledTimes(1);
-      expect(window.gtag).toHaveBeenCalledWith("event", mockEvent, {
-        event_label: mockLabel,
-      });
-    });
-
-    it("should call gtag with value when value given", () => {
-      event(mockEvent, {
-        value: mockValue,
-      });
-
-      expect(window.gtag).toBeCalledTimes(1);
-      expect(window.gtag).toHaveBeenCalledWith("event", mockEvent, {
-        value: mockValue,
-      });
-    });
-  });
 });
 
 describe("pageView", () => {
